changes after contact with clients

+++ lib/app/models/reports/enterprise/enterprise_report_model.dart
@@ -184,7 +184,7 @@ class AmountContacts {
 }
 
 class ValueSold {
-  double valueSoldTotal;
+  int valueSoldTotal;
 
   ValueSold({this.valueSoldTotal});
